+++ r/R/hv.R
@@ -143,9 +143,11 @@ hypervolume <- function(x, reference, maximise = FALSE)
 #' `ignore_dominated` does not change the result, but the default value is
 #' significantly faster.
 #'
-#' The current implementation uses the \eqn{O(n\log n)} dimension-sweep
-#' algorithm for 2D and the naive algorithm that requires calculating the
-#' hypervolume \eqn{|X|+1} times for dimensions larger than 2.
+#' The current implementation uses a \eqn{O(n\log n)} dimension-sweep
+#' algorithm for 2D.  With `ignore_dominated=TRUE`, the 3D case uses the HVC3D
+#' algorithm \citep{GueFon2017hv4d}, which has \eqn{O(n\log n)} complexity.
+#' Otherwise, the implementation uses the naive algorithm that requires
+#' calculating the hypervolume \eqn{|X|+1} times.
 #'
 #' @seealso [hypervolume()]
 #'